detect: move keyword registration into own file
