+++ src/Makefile.am
@@ -206,6 +206,7 @@ detect-ipproto.c detect-ipproto.h \
 detect-iprep.c detect-iprep.h \
 detect-isdataat.c detect-isdataat.h \
 detect-itype.c detect-itype.h \
+detect-keywords.c detect-keywords.h \
 detect-l3proto.c detect-l3proto.h \
 detect-lua.c detect-lua.h \
 detect-lua-extensions.c detect-lua-extensions.h \
