+++ src/tests/detect.c
@@ -23,6 +23,8 @@
 #include "../pkt-var.h"
 #include "../flow-util.h"
 #include "../stream-tcp-reassemble.h"
+#include "../util-unittest.h"
+#include "../util-unittest-helper.h"
 
 static const char *dummy_conf_string =
     "%YAML 1.1\n"
